src/dashboards/ml_dashboard/requirements.txt
streamlit
fastapi
uvicorn
pydantic
pandas
numpy
joblib
requests
PyYAML
Pillow
google-cloud-storage
google-cloud-bigquery
python-dotenv 
db-dtypes
scikit-learn
xgboost

